+++ packages/react-ui/src/index.ts
@@ -77,6 +77,7 @@ import UrlTextField from './shared/UrlTextField';
 import UsernameTextField from './shared/UsernameTextField';
 import EmailTextField from './shared/EmailTextField';
 import PasswordTextField from './shared/PasswordTextField';
+import PhoneTextField from './shared/PhoneTextField';
 import {MEDIA_TYPE_EMBED} from './constants/Media';
 import PollSuggestion, {PollSuggestionProps} from './components/PollSuggestion';
 import ConsentSolution, {ConsentSolutionProps, ConsentSolutionSkeleton} from './components/ConsentSolution';
@@ -95,7 +96,7 @@ import {DEFAULT_WIDGETS_NUMBER} from './constants/Feed';
 import {DEFAULT_PAGINATION_QUERY_PARAM_NAME, DEFAULT_PAGINATION_OFFSET, DEFAULT_PAGINATION_LIMIT} from './constants/Pagination';
 import {FACEBOOK_SHARE, LINKEDIN_SHARE, TWITTER_SHARE} from './constants/SocialShare';
 import {DEFAULT_PRELOAD_OFFSET_VIEWPORT, MAX_PRELOAD_OFFSET_VIEWPORT, MIN_PRELOAD_OFFSET_VIEWPORT} from './constants/LazyLoad';
-import {LEGAL_POLICY_TEC, LEGAL_POLICY_PRIVACY} from './constants/LegalPages';
+import {LEGAL_POLICIES} from './constants/LegalPolicies';
 
 /**
  * Utilities:
@@ -261,6 +262,7 @@ export {
   UsernameTextField,
   EmailTextField,
   PasswordTextField,
+  PhoneTextField,
   MetadataField,
   ConfirmDialog,
   MediaChunkUploader,
@@ -283,8 +285,7 @@ export {
   ConsentSolutionSkeleton,
   ConsentSolutionButton,
   ConsentSolutionButtonProps,
-  LEGAL_POLICY_TEC,
-  LEGAL_POLICY_PRIVACY,
+  LEGAL_POLICIES,
   /* SC UI PAGINATION */
   DEFAULT_PAGINATION_QUERY_PARAM_NAME,
   DEFAULT_PAGINATION_OFFSET,