+++ packages/react-core/CHANGELOG.md
@@ -3,6 +3,25 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.3.8-alpha.7](https://github.com/selfcommunity/community-js/compare/@selfcommunity/react-core@0.3.8-alpha.6...@selfcommunity/react-core@0.3.8-alpha.7) (2022-10-11)
+
+
+### Bug Fixes
+
+* Fix phone number in MetadataField.tsx ([b51a550](https://github.com/selfcommunity/community-js/commit/b51a550dc586c6d90f46f0d195ba0703a57e3967))
+
+
+
+
+
+## [0.3.8-alpha.6](https://github.com/selfcommunity/community-js/compare/@selfcommunity/react-core@0.3.8-alpha.5...@selfcommunity/react-core@0.3.8-alpha.6) (2022-10-11)
+
+**Note:** Version bump only for package @selfcommunity/react-core
+
+
+
+
+
 ## [0.3.8-alpha.5](https://github.com/selfcommunity/community-js/compare/@selfcommunity/react-core@0.3.8-alpha.4...@selfcommunity/react-core@0.3.8-alpha.5) (2022-10-11)